scripts/pylib/robotframework-twister-harness/requirements.txt
# robotframework-twister-harness/requirements.txt
robotframework>=6.0
pytest-twister-harness>=1.0
pyserial>=3.5
psutil>=5.8.0
